added the untire manual gps lifecycle to the app
one problem with it is if your seting the time based on the curretn time - the longitue delta and your using manual gps becasue you dont have gps is you dont have utc time
have to ask the team if they have this some where else on the device.
might have to have them also maually set the time to use - sort of defeating the purpose...

going to alter the layout of the activity to grid or something a bit nicer.